changed int to float

+++ baseclasses/__init__.py
@@ -110,7 +110,7 @@ class BaseProcess(Process):
         a_eln=dict(component='ReferenceEditQuantity'))
 
     positon_in_experimental_plan = Quantity(
-        type=np.dtype(np.int64),
+        type=np.dtype(np.float64),
         a_eln=dict(component='NumberEditQuantity'))
 
     def normalize(self, archive, logger):
